Refactor(step2) : Line Draw 전략에 대한 Refactoring 및 RandomDrawStrategy 클래스 구현, 내부 Magic Number에 대해서는 Refactoring 예정
